[MOSA-149] contracts: updates MockToken.sol and MockToken.js, Mocktoken_utils.js, deletes SimpleToken.js SimpleToken_utils.js  #149
