+++ test/MockToken_utils.js
@@ -22,12 +22,12 @@
 const Assert = require('assert');
 const BigNumber = require('bignumber.js');
 
-var SimpleToken = artifacts.require("./SimpleToken.sol");
+var MockToken = artifacts.require("./MockToken.sol");
 
 /// @dev Deploy 
-module.exports.deploySimpleToken = async (artifacts, accounts) => {
+module.exports.deployMockToken = async (artifacts, accounts) => {
 
-   const token = await SimpleToken.new({ from: accounts[0], gas: 3500000 });
+   const token = await MockToken.new({ from: accounts[0], gas: 3500000 });
 
    return {
       token : token
